+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.11.2
+
+* @iov/cli: Import Lisk, RISE and Ethereum symbols automatically
+* @iov/ethereum: Expose `pubkeyToAddress`
+
 ## 0.11.1
 
 * @iov/ethereum: Fix error reporting when signing and sending transactions